backend/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.32
Werkzeug==3.1.3